Move ActivityQuotedReplyExtention to its own file
